requirements.txt
slack-bolt>=1.18.0
spotipy>=2.23.0
python-dotenv>=1.0.0
fastapi>=0.110.0
uvicorn>=0.27.0
